% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmemory.R
\name{r_code_extract_ggplots_code_list}
\alias{r_code_extract_ggplots_code_list}
\title{Save a history and fix indentation (reformat code in RStudio),
then extract ggplots}
\usage{
r_code_extract_ggplots_code_list(
  script_path = NULL,
  code = readLines(script_path)
)
}
\arguments{
\item{script_path}{a path to a prepared r code script (prepared history)}

\item{code}{r code or path to script}
}
\value{
a list containing ggplot code lines as vectors.
}
\description{
Save a history and fix indentation (reformat code in RStudio),
then extract ggplots
}
\examples{
code = write_ggplots_code_example()
r_code_extract_ggplots_code_list(code = code)
}
